# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/users/orders": {
            "get": {
                "description": "Get Order by OrderId",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "GET Order",
                "parameters": [
                    {
                        "type": "string",
                        "format": "\"Bearer your_actual_token_here\"",
                        "description": "authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "description": "Post a Product to order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Post Order",
                "parameters": [
                    {
                        "type": "string",
                        "format": "\"Bearer your_actual_token_here\"",
                        "description": "authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Order object to be Ordered",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/users/orders/{id}": {
            "get": {
                "description": "Get Order by OrderId",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "GET Order",
                "parameters": [
                    {
                        "type": "string",
                        "format": "\"Bearer your_actual_token_here\"",
                        "description": "authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "\"hexadecimal ObjectId\"",
                        "description": "Order ID to Get from an Order",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entities.Addrs": {
            "type": "object",
            "required": [
                "address",
                "city",
                "postalCode",
                "state"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "postalCode": {
                    "description": "Latitude   float64            `json:\"latitude\" bson:\"latitude\"`\nLongitude  float64            `json:\"longitude\" bson:\"longitude\"`"
                },
                "state": {
                    "type": "string"
                }
            }
        },
...
        "entities.Order": {
            "type": "object",
            "required": [
                "address"
            ],
            "properties": {
                "__v": {
                    "type": "integer"
                },
                "_id": {},
                "address": {
                    "$ref": "#/definitions/entities.Addrs"
                },
                "amountCoupon": {
                    "type": "integer"
                },
                "couponCode": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "isCoupon": {
                    "type": "boolean"
                },
                "jStartDate": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "paymentId": {
                    "type": "string"
                },
                "paymentStatus": {
                    "type": "string"
                },
                "postalCost": {
                    "type": "integer"
                },
                "postalTrakingCode": {
                    "type": "string"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Product"
                    }
                },
                "startDate": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "totalDiscount": {
                    "type": "number"
                },
                "totalPrice": {
                    "type": "integer"
                },
                "totalQuantity": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "entities.Product": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "discountPercent": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "productId": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "variationsKey": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    }
}